core config changes
