+++ grafana-plugin/src/pages/integration_2/Integration2.tsx
@@ -58,7 +58,7 @@ import {
 import { AlertTemplatesDTO } from 'models/alert_templates';
 import { ChannelFilter } from 'models/channel_filter';
 import { MaintenanceType } from 'models/maintenance/maintenance.types';
-import { INTEGRATION_TEMPLATES_LIST, MONACO_PAYLOAD_OPTIONS } from 'pages/integration_2/Integration2.config';
+import { INTEGRATION_TEMPLATES_LIST } from 'pages/integration_2/Integration2.config';
 import IntegrationHelper from 'pages/integration_2/Integration2.helper';
 import styles from 'pages/integration_2/Integration2.module.scss';
 import { PageProps, SelectOption, WithStoreProps } from 'state/types';
@@ -70,6 +70,7 @@ import LocationHelper from 'utils/LocationHelper';
 import { UserActions } from 'utils/authorization';
 import { PLUGIN_ROOT } from 'utils/consts';
 import sanitize from 'utils/sanitize';
+import { MONACO_PAYLOAD_OPTIONS } from './Integration2Common.config';
 
 const cx = cn.bind(styles);
